Removed backends, renamed Docb class to DynamoDB, removed top level all method and moved it to the Query and it is no longer a scan, added sammy as a req
